pyproject.toml
[tool.poetry]
name = "Code Jam 2024"
version = "0.1.0"
description = "The source code of the bot created by team passionate-parsecs during the Python Discord Code Jam 2024"
authors = ["Ajay Ratnam <ajratnam.dev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["voice"], version = "^2.4.0"}
python-dotenv = "^1.0.1"
colorlog = "^6.8.2"
motor = {extras = ["srv"], version = "^3.5.1"}
cogwatch = "^3.3.1"
google-generativeai = "^0.7.2"
beautifulsoup4 = "^4.12.3"
wikipedia = "^1.4.0"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.3"
pre-commit = "^3.7.1"
taskipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python main.py"
lint = "pre-commit run --all-files"
build = "docker build -t code-jam-bot --target=runtime ."
run = "docker run -d code-jam-bot"

[tool.ruff]
# Increase the line length. This breaks PEP8 but it is way easier to work with.
# The original reason for this limit was a standard vim terminal is only 79 characters,
# but this doesn't really apply anymore.
line-length = 119
# Target Python 3.12. If you decide to use a different version of Python
# you will need to update this value.
target-version = "py312"
# Automatically fix auto-fixable issues.
fix = true
# The directory containing the source code. If you choose a different project layout
# you will need to update this value.
src = ["src"]

[tool.ruff.lint]
# Enable all linting rules.
select = ["ALL"]
# Ignore some of the most obnoxious linting errors.
ignore = [
    # Multi Char Strip.
    "B005",
    # Bare Exception.
    "BLE001",
    # Missing docstrings.
    "D100",
    "D104",
    "D105",
    "D106",
    "D107",
    # Docstring whitespace.
    "D203",
    "D213",
    # Docstring punctuation.
    "D415",
    # Docstring quotes.
    "D301",
    # Builtins.
    "A",
    # Print statements.
    "T20",
    # TODOs.
    "TD002",
    "TD003",
    "FIX",
    # Function Definitions.
    "PLR0913",
    # Annotations.
    "ANN101",
    "ANN102",
    "SLF001"
]
